feat(build): initial watch functionality

+++ src/utilities/clean.ts
@@ -1,17 +1,21 @@
 import rimraf = require('rimraf')
-import { outputPath } from '../selectors'
+import { condClean, outputPath } from '../selectors'
 import { store } from '../store'
 
 export const clean = async () =>
   new Promise((resolve, reject) => {
     const state = store.getState()
 
-    rimraf(outputPath(state), err => {
-      // tslint:disable-next-line strict-boolean-expressions
-      if (err) {
-        reject(err)
-      } else {
-        resolve()
-      }
-    })
+    if (condClean(state)) {
+      rimraf(outputPath(state), err => {
+        // tslint:disable-next-line strict-boolean-expressions
+        if (err) {
+          reject(err)
+        } else {
+          resolve()
+        }
+      })
+    } else {
+      resolve()
+    }
   })
